move source code to directory src/
