+++ src/plugins/shs.js
