+++ test/local.js
@@ -1,5 +1,5 @@
 var tape = require('tape')
-var SecretStack = require('../')
+var SecretStack = require('../src')
 var seeds = require('./seeds')
 
 var appkey = Buffer.alloc(32)
